macbook bootstrap: install photo process
Also installs its deps. Improved idempotency

+++ bootstrap/macbook_bootstrap.sh
@@ -17,6 +17,7 @@ APPS=(
   wget
   htop
   syncthing
+  gpsbabel
 )
 
 CASKS=(
@@ -60,8 +61,11 @@ mkdir -p "$HOME"/.local/{bin,src}
 mkdir -p "$HOME"/.config/nvim/bundle
 mkdir -p "$HOME/Downloads/Screenshots"
 
-defaults write com.apple.screencapture location "$HOME/Downloads/Screenshots"
-killall SystemUIServer
+if [[ "$(defaults read com.apple.screencapture location)" != "$HOME/Downloads/Screenshots" ]]; then
+  echo "Setting the default location of screenshots to $HOME/Downloads/Screenshots"
+  defaults write com.apple.screencapture location "$HOME/Downloads/Screenshots"
+  killall SystemUIServer
+fi
 
 # dotfiles!
 cd "$HOME/.local/src" || exit 1
@@ -89,11 +93,27 @@ if [ ! -d "$HOME"/.local/src/autoenv ]; then
   git clone git://github.com/kennethreitz/autoenv.git
 fi
 
+# gpicsync
+if [ ! -d "$HOME/.local/src/GPicSync" ]; then
+  cd "$HOME/.local/src" || exit 1
+  git clone https://github.com/metadirective/GPicSync.git
+fi
+
+cat  > "$HOME/.local/bin/gpicsync" << 'EOF'
+#!/bin/bash
+exec python "$HOME/.local/src/GPicSync/src/gpicsync.py" "$@"
+EOF
+
+chmod +x "$HOME/.local/bin/gpicsync"
+
 if [ ! -f "$HOME"/.local/.bashrc.local ]; then
   echo 'alias vim=nvim' >> "$HOME"/.local/.bashrc.local
 fi
 
 # templates are not really templates yet, so just download them
-wget -O "$HOME"/.config/liquid.theme 'https://github.com/gbagnoli/devmachine/raw/master/site-cookbooks/user/files/default/liquid.theme'
-wget -O "$HOME"/.config/liquidpromptrc 'https://github.com/gbagnoli/devmachine/raw/master/site-cookbooks/user/templates/default/liquidpromptrc.erb'
-wget -O "$HOME"/.bashrc https://github.com/gbagnoli/devmachine/raw/master/site-cookbooks/user/templates/default/bashrc.erb
+wget -nc -O "$HOME"/.config/liquid.theme 'https://github.com/gbagnoli/devmachine/raw/master/site-cookbooks/user/files/default/liquid.theme'
+wget -nc -O "$HOME"/.config/liquidpromptrc 'https://github.com/gbagnoli/devmachine/raw/master/site-cookbooks/user/templates/default/liquidpromptrc.erb'
+wget -nc -O "$HOME"/.bashrc https://github.com/gbagnoli/devmachine/raw/master/site-cookbooks/user/templates/default/bashrc.erb
+
+wget -nc -O "$HOME/.local/bin/photo_process.sh" https://gist.githubusercontent.com/gbagnoli/28565417cfb732cbd2df784819a7fcb0/raw/2e8967cba2d9eb19f9b0fd554b0a91eece075858/photo_process.sh
+chmod +x "$HOME/.local/bin/photo_process.sh"
